DuckDB テストを test_zakuro_duckdb.py に分離する
